Fix Mixed Precision dtype mismatch in Actor/Critic forward passes
Problem: RuntimeError at step 50k - "Index put requires the source and
destination dtypes match, got Float for destination and Half for source"

Root Cause: In Mixed Precision (AMP), forward passes use float16, but
output tensors (q1_vals, q2_vals, actions, log_probs) were initialized
as float32, causing dtype mismatch when writing float16 head outputs.

Solution:
- Critic forward: q1_vals/q2_vals now use trunk_out.dtype (float16 in AMP)
- Actor forward: actions/log_probs now use trunk_out.dtype (float16 in AMP)

Changes:
- mtmh_sac.py:317-318 (Critic): Added dtype=trunk_out.dtype
- mtmh_sac.py:219-220 (Actor): Added dtype=trunk_out.dtype

Verified on DataLAB cluster:
- Job 355814 (MT10): Successfully passed 50k steps (previous failure point)
- No dtype errors
- Mixed Precision Training working correctly

+++ mtmh_sac.py
@@ -214,9 +214,10 @@ class MultiHeadActor(nn.Module):
         
         # Batch processing: group by task_id
         unique_tasks = torch.unique(task_id)
-        
-        actions = torch.zeros((obs.shape[0], self.task_heads[0]['mu'].out_features), device=obs.device)
-        log_probs = torch.zeros((obs.shape[0], 1), device=obs.device) if with_logprob else None
+
+        # Use same dtype as trunk_out (important for Mixed Precision Training)
+        actions = torch.zeros((obs.shape[0], self.task_heads[0]['mu'].out_features), device=obs.device, dtype=trunk_out.dtype)
+        log_probs = torch.zeros((obs.shape[0], 1), device=obs.device, dtype=trunk_out.dtype) if with_logprob else None
         
         for tid in unique_tasks:
             mask = (task_id == tid)
@@ -312,9 +313,10 @@ class MultiHeadCritic(nn.Module):
         
         # Batch processing: group by task_id
         unique_tasks = torch.unique(task_id)
-        
-        q1_vals = torch.zeros(obs.shape[0], device=obs.device)
-        q2_vals = torch.zeros(obs.shape[0], device=obs.device)
+
+        # Use same dtype as trunk_out (important for Mixed Precision Training)
+        q1_vals = torch.zeros(obs.shape[0], device=obs.device, dtype=trunk_out.dtype)
+        q2_vals = torch.zeros(obs.shape[0], device=obs.device, dtype=trunk_out.dtype)
         
         for tid in unique_tasks:
             mask = (task_id == tid)
